hw4/requirements.txt
pytest==8.1.1
javaobj-py3==0.4.4
requests==2.32.5